+++ apps/movies/astro.config.mjs
@@ -2,7 +2,7 @@ import {defineConfig} from 'astro/config'
 
 import react from '@astrojs/react'
 import vercel from '@astrojs/vercel'
-import sanity from '@sanity/astro'
+import sanity from '@avioma/sanity-astro'
 
 // https://astro.build/config
 export default defineConfig({