+++ apps/movies/src/env.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="astro/client" />
-/// <reference types="@sanity/astro/module" />
+/// <reference types="@avioma/sanity-astro/module" />
 
 interface ImportMetaEnv {
   readonly PUBLIC_SANITY_VISUAL_EDITING_ENABLED: string